+++ src/dispatcher/b2bua/invite.rs
@@ -828,6 +828,7 @@ pub fn apply_routing_action(
                 None,
                 None,
                 None,
+                None,
                 &[],
                 state,
             );
@@ -881,6 +882,7 @@ pub fn apply_routing_action(
                     send_socket.as_ref(),
                     None,
                     &message_guard,
+                    None,
                     None,
                     None,
                     None,